+++ src/components/Layout.tsx
@@ -1,14 +1,17 @@
 import React from 'react';
-import { Container, Grid, Stack } from '@mui/material';
+import { Container, Grid } from '@mui/material';
 import { grey } from '@mui/material/colors';
 import { Footer } from './Footer';
 import { Header } from './Header';
 import { Sidebar } from './Sidebar/Sidebar';
 import Breadcrumbs from './Breadcrumbs/Breadcrumbs';
 import { NavigateNext } from '@mui/icons-material';
-import { Outlet } from 'react-router-dom';
+import { Outlet, useMatches } from 'react-router-dom';
+import { CrumbObject } from 'models/Breadcrumbs';
 
 export function Layout() {
+  const matches = useMatches();
+  const crumbs = (matches.find((match) => Boolean(match.handle))?.handle as CrumbObject)?.crumb;
   return (
     <Container maxWidth="lg" disableGutters>
       <Grid container>
@@ -17,9 +20,9 @@ export function Layout() {
         </Grid>
         <Sidebar />
         <Grid item bgcolor={grey['100']} padding={4} xs>
-            <Breadcrumbs separator={<NavigateNext fontSize="small" />} />
+          <Breadcrumbs crumbs={crumbs} separator={<NavigateNext fontSize="small" />} />
 
-            <Outlet />
+          <Outlet />
         </Grid>
         <Grid item xs={12}>
           <Footer />
