+++ src/models/Breadcrumbs.ts
@@ -1,0 +1,16 @@
+export interface BreadcrumbPath {
+  backButton?: boolean;
+  elements: BreadcrumbElement[];
+  routeName: string;
+}
+
+export interface BreadcrumbElement {
+  name: string;
+  fontWeight: number;
+  color?: string;
+  href?: string;
+}
+
+export interface CrumbObject {
+  crumb: BreadcrumbPath;
+}
